+++ packages/@raid-ui/layout/align/src/spacer.js
@@ -2,6 +2,7 @@
 import styled from 'styled-components'
 import * as styledSystem from 'styled-system'
 import systemTypes from '@styled-system/prop-types'
+import propTypes from 'prop-types'
 
 const { compose } = styledSystem
 
@@ -13,11 +14,15 @@ export const Spacer = styled('div')(
     styledSystem.space,
     styledSystem.flexbox,
     styledSystem.layout.display
-  )
+  ),
+  props => ({
+    display: props.isInline ? 'inline-flex' : 'block'
+  })
 )
 Spacer.propTypes = {
   ...systemTypes.space,
   ...systemTypes.flexbox,
-  display: systemTypes.layout.display
+  display: systemTypes.layout.display,
+  isInline: propTypes.bool
 }
 Spacer.displayName = 'Breathe'